Api 파일 폴더로 변경및 기능별 파일 생성
