+++ src/api/auth.js
@@ -1,0 +1,4 @@
+import { API } from "./apiConfig";
+import { WEB_SERVICE } from "../constants/webservice";
+
+export const login = userInfo => API.post("/login", userInfo);
